# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.CreateProductRequest:
    properties:
      category:
        type: string
      description:
        type: string
      image_url:
        type: string
      name:
        type: string
      price:
        minimum: 0
        type: number
      stock:
        minimum: 0
        type: integer
    required:
    - name
    - price
    - stock
    type: object
...
  models.ProductResponse:
    properties:
      category:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      image_url:
        type: string
      name:
        type: string
      price:
        type: number
      stock:
        type: integer
    type: object
...
  models.UpdateProductRequest:
    properties:
      category:
        type: string
      description:
        type: string
      image_url:
        type: string
      name:
        type: string
      price:
        minimum: 0
        type: number
      stock:
        minimum: 0
        type: integer
    type: object
...
  utils.Meta:
    properties:
      filters:
        additionalProperties: true
        type: object
      pagination:
        $ref: '#/definitions/utils.Pagination'
    type: object
  utils.PaginatedResponse:
    properties:
      data: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/utils.Meta'
      status:
        type: integer
    type: object
  utils.Pagination:
    properties:
      current_page:
        type: integer
      has_next:
        type: boolean
      has_prev:
        type: boolean
      items_per_page:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
  utils.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      status:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API backend cho ứng dụng shop bán hàng. Hỗ trợ quản lý sản phẩm,
    xác thực người dùng và quản trị hệ thống.
  title: Backend Shop API
  version: "1.0"
paths:
...
  /products:
    get:
      description: Lấy danh sách sản phẩm với phân trang, tìm kiếm và bộ lọc
      parameters:
      - default: 1
        description: Trang hiện tại
        in: query
        name: page
        type: integer
      - default: 10
        description: Số sản phẩm mỗi trang
        in: query
        maximum: 100
        name: limit
        type: integer
      - description: Từ khóa tìm kiếm
        in: query
        name: search
        type: string
      - description: Danh mục sản phẩm
        in: query
        name: category
        type: string
      - description: Giá tối thiểu
        in: query
        name: min_price
        type: number
      - description: Giá tối đa
        in: query
        name: max_price
        type: number
      - description: Chỉ hiển thị hàng còn
        in: query
        name: in_stock
        type: boolean
      - description: Sắp xếp theo
        enum:
        - price
        - name
        - created_at
        - stock
        - category
        in: query
        name: sort_by
        type: string
      - description: Thứ tự
        enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      - description: Ngày bắt đầu (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: Ngày kết thúc (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.ProductResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Lấy danh sách sản phẩm
      tags:
      - Products
    post:
      consumes:
      - application/json
      description: Tạo sản phẩm mới (chỉ admin)
      parameters:
      - description: Thông tin sản phẩm
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.ProductResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Tạo sản phẩm
      tags:
      - Products
  /products/{id}:
    delete:
      description: Xóa sản phẩm theo ID (chỉ admin)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Xóa sản phẩm
      tags:
      - Products
    get:
      description: Lấy thông tin chi tiết một sản phẩm theo ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.ProductResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Chi tiết sản phẩm
      tags:
      - Products
    put:
      consumes:
      - application/json
      description: Cập nhật thông tin sản phẩm (chỉ admin)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Thông tin sản phẩm
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.ProductResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Cập nhật sản phẩm
      tags:
      - Products
  /products/{id}/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload ảnh cho sản phẩm (chỉ JPG, PNG, GIF)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: File ảnh sản phẩm
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Upload ảnh sản phẩm
      tags:
      - Products
...
securityDefinitions:
  BearerAuth:
    description: 'Nhập token theo format: Bearer {access_token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"